Removed the dependency on commons lang
I consider this library to be fairly low level and would prefer it if
uses as little dependencies as possible.
The danger with these dependencies is that they become available to
clients using light access. Since I consider lang to more or less a
failure and obsolete i prefer to not have it on my classpath at all.